arches/install/requirements_dev.txt
sphinx==1.2.2
sphinx_rtd_theme
fabric
livereload
webtest
django-webtest
sst
django-nose
coverage
sauceclient
